+++ sentry-contexts/build.rs
@@ -16,7 +16,6 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     // See https://github.com/rust-lang/rust/issues/46871
     let arch = target_bits.next().unwrap();
     target_bits.next();
-    let platform = target_bits.next().unwrap();
 
     writeln!(
         f,
@@ -51,11 +50,11 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         }
     )?;
 
-    writeln!(f, "/// The platform identifier.")?;
-    writeln!(
-        f,
-        "#[allow(unused)] pub const PLATFORM: &str = \"{platform}\";"
-    )?;
+    if std::env::var("CARGO_CFG_TARGET_OS").as_deref() == Ok("windows") {
+        writeln!(f, "/// The platform identifier.")?;
+        writeln!(f, "pub const PLATFORM: &str = \"windows\";")?;
+    }
+
     writeln!(f, "/// The CPU architecture identifier.")?;
     writeln!(f, "pub const ARCH: &str = \"{arch}\";")?;